requirements.txt
pygeometa
pyoscar
PyYAML
s3fs
tinydb
